Wine-NSPA: disallow THREAD_PRIORITY_ABOVE_NORMAL as RT. misc.
After a decent amount of testing with the NtCreateThreadEx() RT hacks; I've
determined that ABOVE_NORMAL is not relevant / unnecssary in Wine-NSPA. The
next iteration of NSPA RT support gaurds unix/linux-side main process
threads, so the application is always responsive (as are wine's modules).

   NOTE: I've previously said; the applications' win32 threads will still be run
   at their correct windows process/threads prioclass and priority - and mapped
   to posix/linux. So any non-rt threads will remain SCHED_OTHER.

 RATIONAL:

   - GUI threads are sometimes marked as PROCESS_PRIORITY_REALTIME &&
   THREAD_PRIORIT_ABOVE_NORMAL. While that makes sense in Windows (obviously),
   wine is NOT windows. In Wine-NSPA these threads should NEVER be RT. why;

   - GUI threads run as RT can/do cause xruns.

   - the process itself will now be guarded / high RT priority / wont be
   starved. Additionally, I've yet to see a Windows VST most important
   realtime win32 threads would be using anything lower than
   THREAD_PRIORITY_HIGHEST. TIME_CRITICAL is the norm for critical threads,
   while HIGHEST is often used for some of the processing threads.
   (at least in NI and some other phat VSTi).

   - running the gui thread with FIFO or RR seems like a bad idea.

   non-gui event threads in winelib VST hosts should be PROCESS_PRIOCLASS_REALIME
   and THREAD_PRIORITY_HIGHEST, as to avoid messing with synchronization oor other
   more important threads.

Fix nspa_kernelbase-implement_SetProcessWorkingSetSize.patch. Technically,
I need to disable a check in makedep.c -- which I'm going to ask about on
Wine-devel. This does allow the patch to work, but might not be the sanest
way to do this... although;

technically; I need this hack, if I plan to hook Get/SetThreadPriority()
functions later <= which i plan on doing, so....?

Signed-off-by: ninez <johnstonljordan@gmail.com>